Refactor slice fix gen without unnecessary casting
